chore: refactor ExtractRunModal
